// packages/core/src/types.ts
export type UUID = string;
export type EpochMs = number;

export type PMDoc = {
  type: "doc";
  content: PMNode[];
};

export type PMNode = {
  type: string;
  attrs?: Record<string, any>;
  content?: PMNode[];
  marks?: PMMark[];
  text?: string;
};

export type PMMark = {
  type: string;
  attrs?: Record<string, any>;
};

export type NodeType =
  | "collectionManifest"
  | "collectionMeta"
  | "seriesRef"
  | "series"
  | "seriesOps"
  | "seriesBody"
  | "subseries"
  | "file"
  | "item"
  | "itemFields"
  | "itemBody"
  | "field"
  | "fieldGroup"
  | "suggestion_delete"
  | "suggestion_block"
  | "paragraph"
  | "text"
  | "heading"
  | "bulletList"
  | "orderedList"
  | "listItem"
  | "image";

export type CollectionManifestDoc = {
  type: "doc";
  content: [CollectionManifestNode];
};

export type CollectionManifestNode = {
  type: "collectionManifest";
  attrs: {
    collectionId: UUID;
    title: string;
    dates?: string;
  };
  content: [CollectionMetaNode, ...SeriesRefNode[]];
};

export type CollectionMetaNode = {
  type: "collectionMeta";
  content?: PMNode[];
};

export type SeriesRefNode = {
  type: "seriesRef";
  attrs: {
    seriesId: UUID;
    title: string;
    order: number;
    docName: string;
  };
};

export type SeriesDoc = {
  type: "doc";
  content: [SeriesNode];
};

export type SeriesNode = {
  type: "series";
  attrs: {
    id: UUID;
    title?: string;
    dates?: string;
    refCode?: string;
  };
  content: PMNode[];
};